Remove 'use base' and set @ISA ourself
base.pm is used to import Exporter functionnality in JSON::PP.
But base.pm is an heavy module for what it does.

In recent perl (5.8.3+), we can do "use Exporter 5.57 'import'.
Unfortunately, we still support older perls. So subclassing is the only
supported way.

parent.pm has been created to replace base.pm, but it is available in
core only since 5.10.

So for compatibility with older perls, we subclass Exporter by setting
@ISA directly.