abuse bitwise ops to check for num vs string to avoid B.pm
B.pm is a rather large module to load, and numbers and strings can be
detected by abusing the behavior of bitwise operators.  Using a bitwise
operator is faster as well.

A string & "" results in ""
A number & "" results in 0 (with a warning)

Some variables won't be fully caught by this check alone.  A string like
"0.0" that has been used in a numeric context will pass the check as if
it was a number.  We detect this by making sure the string is the same
as stringifing the numeric form.

"nan" and "inf" that have been used as numbers will also pass that
check, so we detect them by making sure multiplying by 0 gives 0.