Correctly convert objects when TO_JSON returns false
If TO_JSON returns a value that evaluates as false, JSON::PP was
reconverting the object as a value:

    $self->object_to_json($v) || $self->value_to_json($v)

This commit changes the two expressions of that type to dispatch
with a ternary instead:

    ref $v ? $self->object_to_json($v) : $self->value_to_json($v)

As far as I can tell, there were no cases where "object_to_json" was
returning false as a failure condition.  If the undocumented 'fallback'
parameter were changed to do so, then it could happen, but I think
such behavior is a bug.