always encode utf8 flagged scalars as strings
It is very unlikely that a utf8 flagged scalar started as a number.
That would only generally happen if the flag was explicitly enabled.
And that basically counts as modifying the scalar to turn it into a
string.

This also avoids using the bitwise op on a string with wide characters
in it, which is likely to start warning in future perl versions.