add core_bools option to support decoding as core booleans
With this option set, booleans will be decoded into standard perl
boolean values. On older perls, these will not be recognised by the
is_bool function and will not encode as boolean values. On perl
5.35.7 and newer, the values will be able to round trip properly.

Many things assume that the default boolean values are JSON::PP::Boolean
objects, so for now this is not the default.