JSON::PP Generalize for EBCDIC
This commit causes this module to pass all its tests on EBCDIC, which
core Perl supports.  Further, it allows other modules shipped with the
Perl distribution to also work on EBCDIC.

Basically, all the commit does is to remove underlying assumptions of
the bit patterns for various characters, instead using mnemonics, or
functions furnished by core that hide the details.

Some of the tests were converted to use a character, U+00B6, that works
on both character sets, instead of the ones previously used, which
appear to have been arbitrarily chosen.