silence warnings about non-characters on older perls
The last valid unicode codepoint is 0x10FFFF, but this is defined as a
non-character. perl versions <= 5.12 will warn when trying to use the
chr function on strings like this. JSON::PP doesn't enable warnings by
default, but they can still appear when using perl -w, which will often
happen when running tests. These unexpected warnings will break the
tests of ExtUtils::MakeMaker.

Since the module tries to be compatible with prehistoric perl versions,
check the perl version and then disable the utf8 warnings to avoid the
problem.