drop support for perls < 5.8
Support for perl 5.6 and 5.005 has been broken since 4.07, with the use
of utf8::unicode_to_native. While it would be possible to partly fix 5.6
support, there are a variety of things that were already broken, like
incremental parsing of UTF-8 encoded content.

Rather than trying to make those fixes, just drop perl 5.6 support. The
use of 5.6 is extremely small by this point, and properly supporting
anything relating to unicode on 5.6 is very painful.

A number of backward compatibility hacks can be removed. Some of which
weren't ever being used due to their positioning in the file. Remove the
5.6 checks in the tests. And dropping 5.005 allows us to use things like
our.